requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==3.2
django-allauth==0.50.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-heroku==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycountry==22.3.5
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
stripe==2.72.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
